% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\docType{data}
\name{biosemi128_channel_info}
\alias{biosemi128_channel_info}
\title{Channel information for the 128 sensor Biosemi cap}
\format{
A data frame with 128 rows and 9 variables:
\describe{
\item{actiview}{the order in which actiview see the channels}
\item{wired}{the wire connected to each channel}
\item{channel_name}{actual channel name}
\item{channel_name_uppercase}{channel name but all letters are uppercase which can be useful for matching channels as researchers are not always consistent}
\item{X}{channel X coordinate with negative values toward the right ear}
\item{Y}{channel Y coordinate with positive values toward the nose}
\item{Z}{channel Z coordinate with 1 at Cz}
\item{EPosSphere_theta}{angle in radians from Cz north pole to south pole with Cz being 0}
\item{EPosSphere_phi}{angle in radians from right ear (0) toward nose then left ear continuing until back at right ear}
...
}
}
\usage{
biosemi128_channel_info
}
\description{
This data frame details the channels and their positions for the 128 channel
Biosemi cap. These positions are inline with the Oostenveld positions taken
from sphere 1 of his blog (https://robertoostenveld.nl/electrode/#aes1994).
These positions are slightly different than the 64 Biosemi channel cap with
each position being slightly closer together.
}
\keyword{datasets}
