% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\docType{data}
\name{biosemi64_channel_info}
\alias{biosemi64_channel_info}
\title{Channel information for the 64 sensor Biosemi cap}
\format{
A data frame with 64 rows and 9 variables:
\describe{
\item{channel_number}{channel number used in EMEGS and read from AR file}
\item{channel_name}{actual channel name}
\item{channel_name_uppercase}{channel name but all letters are uppercase which can be useful for matching channels as researchers are not always consistent}
\item{X}{channel X coordinate with negative values toward the right ear}
\item{Y}{channel Y coordinate with positive values toward the nose}
\item{Z}{channel Z coordinate with 1 at Cz}
\item{EPosSphere_theta}{angle in radians from Cz north pole to south pole with Cz being 0}
\item{EPosSphere_phi}{angle in radians from right ear (0) toward nose then left ear continuing until back at right ear}
...
}
}
\usage{
biosemi64_channel_info
}
\description{
This data frame shows the channel names in order as well as each channels
Cartesian and Spherical coordinates. You may notice that the channels are
slightly lower than the 128 channels or the Oostenveld channels. This is
because this is the way the Biosemi the company and an older version of
EMEGS thought the channels were located. It is a small difference, but we
leaving the older channel positions for this cap such that it is consistent
with our already published studies. Cartesian coordinates assume a radius of
1.
}
\keyword{datasets}
