% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\docType{data}
\name{biosemi64_channel_numbers_names}
\alias{biosemi64_channel_numbers_names}
\title{Channel names and numbers for the biosemi 64 channel cap}
\format{
A data frame with 65 rows and 2 variables:
\describe{
\item{channel_number}{channel number used in EMEGS and read from AR file}
\item{channel_name}{actual channel name}
...
}
}
\usage{
biosemi64_channel_numbers_names
}
\description{
This small data frame can help to show the name for each channel number.
}
\keyword{datasets}
