% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{cart2phi_degrees}
\alias{cart2phi_degrees}
\title{Finds phi in degrees with a X and Y coordinates}
\usage{
cart2phi_degrees(x_coor, y_coor)
}
\arguments{
\item{x_coor}{x coordinate between -1 and 1}

\item{y_coor}{y coordinate between -1 and 1}
}
\description{
Takes Cartesian X and Y coordinates and returns phi angle in degrees which is 0
at the right ear or an X at -1 and Y at 0.
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
