% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{cart2polar}
\alias{cart2polar}
\title{Turns Cartesian coordinates into polar coordinates}
\usage{
cart2polar(x_coor, y_coor, z_coor)
}
\arguments{
\item{x_coor}{x coordinate between -1 and 1}

\item{y_coor}{y coordinate between -1 and 1}

\item{z_coor}{z coordinate between -1 and 1}
}
\value{
Returns a named vector with theta and phi in degrees and radians
}
\description{
This function takes Cartesian coordinates and turns them into polar/geographic
coordinates. Finding these positions are necessary for creating ecfg configuration files for EMEGS
such that the software knows its three dimensional location on a sphere. Assumes
radius of one, X is right ear (negative) to left ear (positive), Y positive toward nose
, and Z is 1 at Cz. Spherical coordinates are returned in radians and degrees. Theta is 0 at Cz
pi at other pole. Phi is zero at right ear, pi/2 (90°) at nose, pi (180°) at left ear, 3/2*pi (270°) at
Oz.
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
