% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{cart2theta_radians}
\alias{cart2theta_radians}
\title{Finds theta in radians with a Z coordinate}
\usage{
cart2theta_radians(z_coor)
}
\arguments{
\item{z_coor}{z coordinate between -1 and 1}
}
\description{
Takes Cartesian Z coordinate and returns theta angle in radians which is 0
at Cz which is a Z coordinate of 1
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
