% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{deg2rad}
\alias{deg2rad}
\title{Turns degrees into radians}
\usage{
deg2rad(degrees)
}
\arguments{
\item{degrees}{the degrees number to be turned to degrees}
}
\description{
Simple function that turns degrees to radians.
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
