% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{get_original_markers}
\alias{get_original_markers}
\title{Replaces marker file with the original}
\usage{
get_original_markers(folders = NULL)
}
\arguments{
\item{folders}{select at least one or more data folders to search for marker files}
}
\description{
This function replaced the edited marker file with the original
}
\author{
Andrew H Farkas, \email{andrewhfarkas@gmail.com}
}
