% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{marker_file_editor}
\alias{marker_file_editor}
\title{Make changes to ctf MEG marker file and save the original}
\usage{
marker_file_editor(
  folders = NULL,
  target_triggers = NULL,
  delete_trigger_numbers = NULL
)
}
\arguments{
\item{folders}{select at least one or more data folders to search for marker files}

\item{target_triggers}{at least one or more target triggers}

\item{delete_trigger_numbers}{numeric vector of which triggers should be removed
from marker file}
}
\description{
This function saves original marker files used in ctf MEG files (.ds files)
new folder called original_marker_files in the parent folder. It then will
alter one present in the folder such that triggers can be removed or adjusted.
}
\author{
Andrew H Farkas, \email{andrewhfarkas@gmail.com}
}
