% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\docType{data}
\name{Oostenveld_channel_info}
\alias{Oostenveld_channel_info}
\title{Channel information for all positions courtesy of Oostenveld's blog}
\format{
A data frame with 128 rows and 7 variables:
\describe{
\item{channel_name}{actual channel name}
\item{channel_name_uppercase}{channel name but all letters are uppercase which can be useful for matching channels as researchers are not always consistent}
\item{X}{channel X coordinate with negative values toward the right ear}
\item{Y}{channel Y coordinate with positive values toward the nose}
\item{Z}{channel Z coordinate with 1 at Cz}
\item{EPosSphere_theta}{angle in radians from Cz north pole to south pole with Cz being 0}
\item{EPosSphere_phi}{angle in radians from right ear (0) toward nose then left ear continuing until back at right ear}
...
}
}
\usage{
Oostenveld_channel_info
}
\description{
This data frame details the channels and their positions for the positions
outlined by Dr. Oostenveld. These positions where taken from Oostenveld's
sphere 1 off of his blog (https://robertoostenveld.nl/electrode/#aes1994).
These positions are slightly different than the 64 Biosemi channel cap with
each position being slightly closer together. This was used to find the
positions for the 128 biosemi cap
}
\keyword{datasets}
