% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{polar2cart}
\alias{polar2cart}
\title{Turns polar coordinates (EPosSphere) into Cartesian coordinates}
\usage{
polar2cart(theta_radians, phi_radians)
}
\arguments{
\item{theta_radians}{angle in radians from north pole Cz}

\item{phi_radians}{angle in radians from right ear to nose}
}
\value{
Returns a named vector with X, Y, and Z coordinates
}
\description{
This function takes polar coordinates (EPosSphere in EMEGS) and turns them into
Cartesian 3d coordinates. Assumes radius of one, X is right ear (negative) to left ear (positive),
Y positive toward nose, and Z is 1 at Cz. Spherical coordinates are returned in radians
and degrees. Theta is 0 at Cz pi at other pole. Phi is zero at right ear, pi/2 (90°) at
nose, pi (180°) at left ear, 3/2*pi (270°) at Oz.
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
