% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{rad2deg}
\alias{rad2deg}
\title{Turns radians into degrees}
\usage{
rad2deg(radians)
}
\arguments{
\item{radians}{the radians number to be turned to degrees}
}
\description{
Simple function that turns radians to degrees.
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
