% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{read_ar_file}
\alias{read_ar_file}
\title{Read ERP data from one AR file}
\usage{
read_ar_file(path_to_ar = NULL, extract_channels = NULL, baseline_pts = NULL)
}
\arguments{
\item{path_to_ar}{The path to the AR file}

\item{extract_channels}{Choose which channels to extract, must know numeric
index (can be found in Emegs2d)}

\item{baseline_pts}{Select baseline points to baseline the channels as they
read in. Ex: c(0:50) which is the first 51 data points}
}
\value{
Returns a dataframe of the ERP data extracted from an AR file
}
\description{
This function reads data from one AR file that come from EMEGS.
}
\author{
Andrew H Farkas, \email{andrewhfarkas@gmail.com}
}
