% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{read_ar_files}
\alias{read_ar_files}
\title{Read ERP data from multiple AR files}
\usage{
read_ar_files(
  data_folders = NULL,
  extract_channels = NULL,
  baseline_pts = NULL,
  patterns = NULL,
  search_subfolders = F,
  search_only_subfolders = F
)
}
\arguments{
\item{data_folders}{select multiple data folders to search}

\item{extract_channels}{Choose which channels to extract, must know numeric
index (can be found in Emegs2d)}

\item{baseline_pts}{Select baseline points to baseline the channels as they
read in. Ex: c(0:50) which is the first 51 data points}

\item{patterns}{One or more patterns to find files}

\item{search_only_subfolders}{Argument to specify if the files in the parent
folder should not be searched}

\item{search_subfolder}{Logical argument to specify if subfolder should be
searched}
}
\value{
Returns a dataframe of the ERP data extracted from AR files
}
\description{
This function reads data from AR files that come from EMEGS.
}
\author{
Andrew H Farkas, \email{andrewhfarkas@gmail.com}
}
