% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{se}
\alias{se}
\title{Finds the standard error}
\usage{
se(vec, na.rm = FALSE)
}
\arguments{
\item{vec}{a vector of numerical values}

\item{na.rm}{a logical argument to remove NA values before calculating the
standard error}
}
\value{
Returns a numeric standard error value
}
\description{
A simple function to find standard error. There is different defintions of
standard error so beware
}
\author{
Andrew H Farkas, \email{andrewhfarkas at g mail dot com}
}
