% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{windows_emegs_batchfile_maker}
\alias{windows_emegs_batchfile_maker}
\title{Create an EMEGS batchfile for a windows directory system}
\usage{
windows_emegs_batchfile_maker(
  patterns,
  batch_path,
  parent_path,
  search_subfolders = F,
  search_only_subfolders = F,
  patterns_exclude
)
}
\arguments{
\item{patterns}{1 or more patterns to find files}

\item{batch_path}{Where the batch text file will be written, must
include filename with .txt ending if specifying path}

\item{parent_path}{Path to where the function will start searching}

\item{search_subfolders}{Argument to specify if the function should search
folders in the parent_path}

\item{search_only_subfolders}{Argument to specify if the files in the parent
folder should not be searched}

\item{patterns_exclude}{not currently in use}
}
\value{
writes a batch text file in batch_path
}
\description{
This function uses r regex patterns to find files and writes
batch text files which EMEGS can read. It is capable of searching
in subfolders. This is for the windows directory system. A mac
version (if finished) is called mac_emegs_batchfile_maker.
}
\author{
Andrew H Farkas, \email{andrewhfarkas@gmail.com}
}
