% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEGShelper.R
\name{read_ar_file}
\alias{read_ar_file}
\title{Read ERP data from one AR file}
\usage{
read_ar_file(
  path_to_ar = NULL,
  extract_channels = NULL,
  baseline_pts = NULL,
  select_time_points = NULL,
  average_timepoints = F,
  average_channels = F,
  include_channel_name = F,
  include_path_name = F,
  include_file_name = F
)
}
\arguments{
\item{path_to_ar}{The path to the AR file}

\item{extract_channels}{Choose which channels to extract, must know numeric
index (can be found in Emegs2d)}

\item{baseline_pts}{Select baseline points to baseline the channels as they
read in. Ex: c(0:50) which is the first 51 data points}

\item{select_time_points}{Select time points to keep. Works after baselining. You select sample
points not in millisecond. You will need to know your samplerate and scene onset}

\item{average_timepoints}{Logical argument if selected time points should be averaged}

\item{average_channels}{Logical argument to average selected channels together}

\item{include_channel_name}{Logical argument for whether to put channel names in output dataframe}

\item{include_path_name}{Logical argument for whether to put path name in output dataframe}

\item{include_file_name}{Logical argument for whether to put file name in output dataframe}
}
\value{
Returns a dataframe of the ERP data extracted from an AR file
}
\description{
This function reads data from one AR file that come from EMEGS.
}
\author{
Andrew H Farkas, \email{andrewhfarkas@gmail.com}
}
